32 bit

+++ src/mcu.h
@@ -291,7 +291,7 @@ struct MCU {
 
     int rom2_mask = ROM2_SIZE - 1;
 
-    short sample_buffer[audio_buffer_size] = {0};
+    int sample_buffer[audio_buffer_size] = {0};
     int sample_write_ptr = 0;
 
     int ga_int[8] = {0};
@@ -367,7 +367,7 @@ struct MCU {
     void MCU_EncoderTrigger(int dir);
 
     int startSC55(std::string *basepath);
-    int updateSC55(int16_t *data, unsigned int dataSize);
+    int updateSC55(int32_t *data, unsigned int dataSize);
     void postMidiSC55(uint8_t* message, int length);
     void SC55_Reset();
